Demo_AI/ai-command-assistant/requirements.txt
# 确保有httpx库，DeepSeek API需要
httpx>=0.25.0
python-dotenv>=1.0.0